Refactors directory structure, notably eventhubs, not eventhub
The name of the Azure service is "Event Hubs", not Eventhub or similar.
I've renamed package and config names to match Microsoft's convention,
which is interestingly inconsistent between services. For example,
"IoT Hub" and "Event Hubs"

I also removed the "module" package in favor of collector. "module" is
a side-effect of the autoconfiguration build.

Finally, I cleaned up some sentences.